// web/lib/types/WorkloadLogsResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LogLine } from "./LogLine";

export type WorkloadLogsResponse = { workload_id: string, host_id: string, fetched_at: string, lines: Array<LogLine>, truncated: boolean, error: string | null, };